+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Codeception v2.0.11 support.
 - Fixed: uncaught exception when a site has a suite «foo», and the other site does not have this suite.
+- Added option «debug» which toggles Codeception debug messages.
+- Added option «colors» which toggles colors on Webception Console.
 
 ## 0.1.0 (2014-01-07)